# docs/api/swagger.yaml
definitions:
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion:
    properties:
      amount:
        type: number
      created:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType'
      updated:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType:
    enum:
    - regular
    - welcome_bonus
    type: string
    x-enum-varnames:
    - Regular
    - WelcomeBonus
...
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User:
    properties:
      balance:
        type: number
      created:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      promotions:
        items:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
        type: array
      role:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserType'
      updated:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion:
    properties:
      claimed:
        type: string
      created:
        type: string
      end_date:
        type: string
      id:
        type: string
      promotion:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
      promotion_id:
        type: string
      start_date:
        type: string
      updated:
        type: string
      user:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
      user_id:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Player
    - Staff
...
info:
  contact: {}
paths:
...
  /api/v1/user-promotions/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all promotions assigned to a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user promotions
          schema:
            items:
              $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
            type: array
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get all promotions for a user
      tags:
      - User Promotions
    post:
      consumes:
      - application/json
      description: Assign a promotion to a user with the provided details
      parameters:
      - description: User Promotion details
        in: body
        name: userPromotion
        required: true
        schema:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
      produces:
      - application/json
      responses:
        "200":
          description: Added user promotion
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
        "400":
          description: Invalid input or business rule violation
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Add a promotion to a user
      tags:
      - User Promotions
  /api/v1/user-promotions/{user_id}/promotion/{user_prom_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user promotion using its unique ID
      parameters:
      - description: User Promotion ID
        in: path
        name: user_prom_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved user promotion
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "404":
          description: User promotion not found
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get a user promotion by ID
      tags:
      - User Promotions
  /api/v1/user-promotions/{user_id}/promotions/{user_prom_id}:
    delete:
      consumes:
      - application/json
      description: Delete a user promotion using its unique ID
      parameters:
      - description: User Promotion ID
        in: path
        name: user_prom_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Delete a user promotion
      tags:
      - User Promotions
  /api/v1/user-promotions/{user_id}/promotions/{user_prom_id}/claim:
    post:
      consumes:
      - application/json
      description: Allows a user to claim a promotion if eligible
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: User Promotion ID
        in: path
        name: user_prom_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid input or business rule violation
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "403":
          description: Forbidden - Requestor ID does not match
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Claim a promotion
      tags:
      - User Promotions
...
swagger: "2.0"